requirements.txt
SQLAlchemy~=1.4.41
fastapi~=0.85.0
zipstream~=1.1.4
pydantic~=1.10.2
noid-mint~=0.0.3